Refactored changes to comply with coding standard and put the API behind an interface to make future testing and development less coupled to the implementation.
