+++ src/HackF5.UnitySpy.HearthstoneLib/Detail/HearthstoneImage.cs
@@ -1,31 +1,30 @@
-﻿namespace HackF5.UnitySpy.HearthstoneLib
+﻿namespace HackF5.UnitySpy.HearthstoneLib.Detail
 {
-    using System;
-
-    public class HearthstoneImage
+    internal class HearthstoneImage
     {
-        private IAssemblyImage image;
+        private readonly IAssemblyImage image;
 
         public HearthstoneImage(IAssemblyImage image)
         {
             this.image = image;
         }
 
-        public dynamic this[string fullTypeName] { 
-            get {
-                return image[fullTypeName];
-            }
-        }
+        public dynamic this[string fullTypeName] => this.image[fullTypeName];
 
         public dynamic GetService(string name)
         {
-            var dynamicServices = image?["HearthstoneServices"]["s_dynamicServices"];
-            var staticServices = image?["HearthstoneServices"]["s_services"];
-            var services = dynamicServices != null ? dynamicServices : staticServices;
+            var dynamicServices = this.image?["HearthstoneServices"]["s_dynamicServices"];
+            var staticServices = this.image?["HearthstoneServices"]["s_services"];
+            var services = dynamicServices ?? staticServices;
+
+            if (services == null)
+            {
+                return null;
+            }
 
             var serviceItems = services["m_services"]["_items"];
 
-            int i = 0; ;
+            var i = 0;
             foreach (var service in serviceItems)
             {
                 if (service?["<ServiceTypeName>k__BackingField"] == name)
@@ -33,6 +32,7 @@
                     var result = service["<Service>k__BackingField"];
                     return result;
                 }
+
                 i++;
             }
