+++ src/HackF5.UnitySpy.HearthstoneLib/Detail/Match/Account.cs
@@ -1,0 +1,9 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib.Detail.Match
+{
+    internal class Account : IAccount
+    {
+        public ulong Hi { get; set; }
+
+        public ulong Lo { get; set; }
+    }
+}
