+++ src/HackF5.UnitySpy.HearthstoneLib/Detail/Match/BattleTag.cs
@@ -1,0 +1,9 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib.Detail.Match
+{
+    internal class BattleTag : IBattleTag
+    {
+        public string Name { get; set; }
+
+        public int Number { get; set; }
+    }
+}
