+++ src/HackF5.UnitySpy.HearthstoneLib/IAccount.cs
@@ -1,0 +1,12 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib
+{
+    using JetBrains.Annotations;
+
+    [PublicAPI]
+    public interface IAccount
+    {
+        ulong Hi { get; set; }
+
+        ulong Lo { get; set; }
+    }
+}
