+++ src/HackF5.UnitySpy.HearthstoneLib/IBattleTag.cs
@@ -1,0 +1,12 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib
+{
+    using JetBrains.Annotations;
+
+    [PublicAPI]
+    public interface IBattleTag
+    {
+        string Name { get; }
+
+        int Number { get; }
+    }
+}
