+++ src/HackF5.UnitySpy.HearthstoneLib/ICollectionCard.cs
@@ -1,0 +1,14 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib
+{
+    using JetBrains.Annotations;
+
+    [PublicAPI]
+    public interface ICollectionCard
+    {
+        string CardId { get; }
+
+        int Count { get; }
+
+        int PremiumCount { get; }
+    }
+}
