--- src/HackF5.UnitySpy.HearthstoneLib/Match/Account.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace HackF5.UnitySpy.HearthstoneLib.Match
-{
-    public class Account
-    {
-        public ulong Hi { get; set; }
-        public ulong Lo { get; set; }
-    }
-}
