--- src/HackF5.UnitySpy.HearthstoneLib/Match/BattleTag.cs
@@ -1,8 +1,0 @@
-﻿namespace HackF5.UnitySpy.HearthstoneLib.Match
-{
-    public class BattleTag
-    {
-        public string Name { get; set; }
-        public int Number { get; set; }
-    }
-}
