--- src/HackF5.UnitySpy.HearthstoneLib/Match/Side.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace HackF5.UnitySpy.HearthstoneLib.Match
-{
-    public enum Side
-    {
-        NEUTRAL,
-        FRIENDLY,
-        OPPOSING
-    }
-}
