+++ src/HackF5.UnitySpy.HearthstoneLib/Side.cs
@@ -1,0 +1,12 @@
+﻿namespace HackF5.UnitySpy.HearthstoneLib
+{
+    using JetBrains.Annotations;
+
+    [PublicAPI]
+    public enum Side
+    {
+        NEUTRAL,
+        FRIENDLY,
+        OPPOSING,
+    }
+}
