+++ src/HackF5.UnitySpy.Tests/TestHearthstoneLibApi.cs
@@ -1,21 +1,23 @@
-
-
 namespace HackF5.UnitySpy.HearthstoneLib.Tests
 {
+    using JetBrains.Annotations;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using System;
 
     // This needs to be run while Hearthstone is running
     [TestClass]
     public class TestHearthstoneLibApi
     {
+        [PublicAPI]
+        public TestContext TestContext { get; set; }
+
         [TestMethod]
         public void TestRetrieveCollection()
         {
             var collection = new MindVision().GetCollection();
             Assert.IsNotNull(collection);
-            Assert.IsTrue(collection.Count > 0, "collection should not be empty");
-            Console.WriteLine("Collection had " + collection.Count + " cards");
+            Assert.IsTrue(collection.Count > 0, "Collection should not be empty.");
+
+            this.TestContext.WriteLine($"Collection has {collection.Count} cards.");
         }
 
         // You need to have a game running for this
@@ -24,7 +26,7 @@ namespace HackF5.UnitySpy.HearthstoneLib.Tests
         {
             var matchInfo = new MindVision().GetMatchInfo();
             Assert.IsNotNull(matchInfo);
-            Console.WriteLine(matchInfo.LocalPlayer.StandardRank);
+            this.TestContext.WriteLine($"Local player's standard rank is {matchInfo.LocalPlayer.StandardRank}.");
         }
     }
 }
